Extract Community "about" tab to separate view
