+++ Mlem/App/Views/Pages/Community/CommunityView+Logic.swift
@@ -18,11 +18,6 @@ extension CommunityView {
         return firstPerson.isAdmin || firstPerson.moderates(community: community)
     }
 
-    func canEditDescription(_ community: any Community) -> Bool {
-        guard let firstPerson = appState.firstPerson else { return false }
-        return firstPerson.isAdmin || firstPerson.moderates(community: community)
-    }
-    
     func openAddModSheet() {
         navigation.openSheet(.personPicker { person in
             newMod = person
